Checking in  sd card dll and other interface changes

+++ MicroFrameworkPK_v4_3/DeviceCode/Targets/Native/STM32F10x/DeviceCode/drivers/Samraksh_SPOT_Hardware/spot_native_Samraksh_SPOT_Hardware_EmoteDotNow_NOR.h
@@ -27,9 +27,9 @@ namespace Samraksh
                     // Helper Functions to access fields of managed object
                     // Declaration of stubs. These functions are implemented by Interop code developers
                     static INT8 InternalInitialize( HRESULT &hr );
-                    static INT8 InternalWrite( CLR_RT_TypedArray_UINT16 param0, UINT32 param1, UINT16 param2, HRESULT &hr );
-                    static INT8 InternalWrite( CLR_RT_TypedArray_UINT16 param0, UINT32 param1, UINT16 param2, UINT16 param3, HRESULT &hr );
-                    static INT8 InternalRead( CLR_RT_TypedArray_UINT16 param0, UINT32 param1, UINT16 param2, HRESULT &hr );
+                    static INT32 InternalWrite( CLR_RT_TypedArray_UINT16 param0, UINT32 param1, UINT16 param2, HRESULT &hr );
+                    static INT32 InternalWrite( CLR_RT_TypedArray_UINT16 param0, UINT32 param1, UINT16 param2, UINT16 param3, HRESULT &hr );
+                    static INT32 InternalRead( CLR_RT_TypedArray_UINT16 param0, UINT32 param1, UINT16 param2, HRESULT &hr );
                 };
             }
         }
