+++ MicroFrameworkPK_v4_3/Samraksh/Interops/Samraksh_SPOT_Hardware/Samraksh_SPOT_Hardware/NativeStubs/spot_native_Samraksh_SPOT_Hardware_EmoteDotNow_NOR.cpp
@@ -22,21 +22,21 @@ INT8 NOR::InternalInitialize( HRESULT &hr )
     return retVal;
 }
 
-INT8 NOR::InternalWrite( CLR_RT_TypedArray_UINT16 param0, UINT32 param1, UINT16 param2, HRESULT &hr )
+INT32 NOR::InternalWrite( CLR_RT_TypedArray_UINT16 param0, UINT32 param1, UINT16 param2, HRESULT &hr )
 {
-    INT8 retVal = 0; 
+    INT32 retVal = 0; 
     return retVal;
 }
 
-INT8 NOR::InternalWrite( CLR_RT_TypedArray_UINT16 param0, UINT32 param1, UINT16 param2, UINT16 param3, HRESULT &hr )
+INT32 NOR::InternalWrite( CLR_RT_TypedArray_UINT16 param0, UINT32 param1, UINT16 param2, UINT16 param3, HRESULT &hr )
 {
-    INT8 retVal = 0; 
+    INT32 retVal = 0; 
     return retVal;
 }
 
-INT8 NOR::InternalRead( CLR_RT_TypedArray_UINT16 param0, UINT32 param1, UINT16 param2, HRESULT &hr )
+INT32 NOR::InternalRead( CLR_RT_TypedArray_UINT16 param0, UINT32 param1, UINT16 param2, HRESULT &hr )
 {
-    INT8 retVal = 0; 
+    INT32 retVal = 0; 
     return retVal;
 }
